+++ assembly-syntax/src/ast/visit.rs
@@ -322,7 +322,7 @@ where
         | MemLoadWImm(imm)
         | MemStoreImm(imm)
         | MemStoreWImm(imm)
-        | Emit(imm)
+        | EmitImm(imm)
         | Trace(imm) => visitor.visit_immediate_u32(imm),
         SysEvent(sys_event) => visitor.visit_system_event(Span::new(span, sys_event)),
         Exec(target) => visitor.visit_exec(target),
@@ -350,7 +350,7 @@ where
         | Sdepth | Caller | Clk | MemLoad | MemLoadW | MemStore | MemStoreW | MemStream
         | AdvPipe | AdvLoadW | Hash | HMerge | HPerm | MTreeGet | MTreeSet | MTreeMerge
         | MTreeVerify | FriExt2Fold4 | DynExec | DynCall | Breakpoint | HornerBase | HornerExt
-        | EvalCircuit => ControlFlow::Continue(()),
+        | EvalCircuit | Emit => ControlFlow::Continue(()),
     }
 }
 
@@ -753,7 +753,7 @@ where
         | MemLoadWImm(imm)
         | MemStoreImm(imm)
         | MemStoreWImm(imm)
-        | Emit(imm)
+        | EmitImm(imm)
         | Trace(imm) => visitor.visit_mut_immediate_u32(imm),
         SysEvent(sys_event) => visitor.visit_mut_system_event(Span::new(span, sys_event)),
         Exec(target) => visitor.visit_mut_exec(target),
@@ -781,7 +781,7 @@ where
         | Sdepth | Caller | Clk | MemLoad | MemLoadW | MemStore | MemStoreW | MemStream
         | AdvPipe | AdvLoadW | Hash | HMerge | HPerm | MTreeGet | MTreeSet | MTreeMerge
         | MTreeVerify | FriExt2Fold4 | DynExec | DynCall | Breakpoint | HornerBase | HornerExt
-        | EvalCircuit => ControlFlow::Continue(()),
+        | EvalCircuit | Emit => ControlFlow::Continue(()),
     }
 }