+++ assembly/src/basic_block_builder.rs
@@ -98,7 +98,8 @@ impl BasicBlockBuilder<'_> {
     /// Converts the system event into its corresponding event ID, and adds an `Emit` operation
     /// to the list of basic block operations.
     pub fn push_system_event(&mut self, sys_event: SystemEvent) {
-        self.push_op(Operation::Emit(sys_event.into_event_id()))
+        let event_id = sys_event.into_event_id();
+        self.push_ops([Operation::Push(Felt::from(event_id)), Operation::Emit, Operation::Drop]);
     }
 }