+++ core/src/operations/mod.rs
@@ -167,9 +167,10 @@ pub enum Operation {
     /// - 16 most significant bits identify the event source,
     /// - 16 least significant bits identify the actual event.
     ///
-    /// Similar to Noop, this operation does not change the state of user stack. The immediate
-    /// value affects the program MAST root computation.
-    Emit(u32) = OPCODE_EMIT,
+    /// This operation does not change the state of the user stack. The event ID is read
+    /// from the top of the stack without consuming it. The event id must be manually
+    /// pushed and dropped.
+    Emit = OPCODE_EMIT,
 
     // ----- flow control operations -------------------------------------------------------------
     /// Marks the beginning of a join block.
@@ -622,7 +623,6 @@ impl Operation {
     pub fn imm_value(&self) -> Option<Felt> {
         match *self {
             Self::Push(imm) => Some(imm),
-            Self::Emit(imm) => Some(imm.into()),
             _ => None,
         }
     }
@@ -771,7 +771,7 @@ impl fmt::Display for Operation {
             Self::MStream => write!(f, "mstream"),
             Self::Pipe => write!(f, "pipe"),
 
-            Self::Emit(value) => write!(f, "emit({value})"),
+            Self::Emit => write!(f, "emit"),
 
             // ----- cryptographic operations -----------------------------------------------------
             Self::HPerm => write!(f, "hperm"),
@@ -799,7 +799,6 @@ impl Serializable for Operation {
                 err_code.write_into(target);
             },
             Operation::Push(value) => value.as_int().write_into(target),
-            Operation::Emit(value) => value.write_into(target),
 
             // Note: we explicitly write out all the operations so that whenever we make a
             // modification to the `Operation` enum, we get a compile error here. This
@@ -810,6 +809,7 @@ impl Serializable for Operation {
             | Operation::SDepth
             | Operation::Caller
             | Operation::Clk
+            | Operation::Emit
             | Operation::Join
             | Operation::Split
             | Operation::Loop
@@ -993,11 +993,7 @@ impl Deserializable for Operation {
 
             OPCODE_MRUPDATE => Self::MrUpdate,
             OPCODE_PUSH => Self::Push(Felt::read_from(source)?),
-            OPCODE_EMIT => {
-                let value = source.read_u32()?;
-
-                Self::Emit(value)
-            },
+            OPCODE_EMIT => Self::Emit,
             OPCODE_SYSCALL => Self::SysCall,
             OPCODE_CALL => Self::Call,
             OPCODE_END => Self::End,