+++ core/src/sys_events.rs
@@ -7,6 +7,8 @@ use core::fmt;
 // between 0 and 2^32.
 pub use constants::*;
 
+use crate::Felt;
+
 #[rustfmt::skip]
 mod constants {
     pub const EVENT_MERKLE_NODE_MERGE: u32            = 276124218;
@@ -318,7 +320,8 @@ impl SystemEvent {
 
     /// Returns a system event corresponding to the specified event ID, or `None` if the event
     /// ID is not recognized.
-    pub fn from_event_id(event_id: u32) -> Option<Self> {
+    pub fn from_event_id(event_id: Felt) -> Option<Self> {
+        let event_id: u32 = event_id.as_int().try_into().ok()?;
         match event_id {
             EVENT_MERKLE_NODE_MERGE => Some(SystemEvent::MerkleNodeMerge),
             EVENT_MERKLE_NODE_TO_STACK => Some(SystemEvent::MerkleNodeToStack),